Implement purging
- switched from `pip` and `subprocess` to `importlib.metadata`

- also did some restructure
