+++ pip_manage/pip_purge.py
@@ -11,26 +11,15 @@ from pip_manage._logging import setup_logging
 from pip_manage._pip_interface import (
     filter_forwards,
     get_dependencies_of_package,
-    purge_packages,
+    uninstall_packages,
 )
 
 if TYPE_CHECKING:
     import logging
     from collections.abc import Sequence
 
-# parameters that pip uninstall supports
-_UNINSTALL_ONLY: Final[frozenset[str]] = frozenset(
-    (
-        "r",
-        "requirement",
-        "y",
-        "yes",
-        "root-user-action",
-        "break-system-packages",
-    ),
-)
 
-
+# TODO: Add 'exclude' argument
 def _parse_args(
     args: Sequence[str] | None = None,
 ) -> tuple[argparse.Namespace, list[str]]:
@@ -52,8 +41,9 @@ def _parse_args(
     return parser.parse_known_args(args)
 
 
-# TODO: Implement purging
-# TODO: Improve naming
+# TODO: Improve naming!!!
+# TODO: Ask for conformation
+# TODO: Actually use 'filter_forwards'
 def main(argv: Sequence[str] | None = None) -> int:
     args, forwarded = _parse_args(argv)
     uninstall_args: list[str] = filter_forwards(forwarded, set())
@@ -64,22 +54,39 @@ def main(argv: Sequence[str] | None = None) -> int:
     dct: dict[str, tuple[frozenset[str], frozenset[str]]] = {}
     for pkg in args.pkgs:
         dct[pkg] = get_dependencies_of_package(pkg)
-    logger.debug(dct)
+        for pkg_depd in dct[pkg][0]:
+            dct[pkg_depd] = get_dependencies_of_package(pkg_depd)
 
+    pkgs_to_not_be_uninstalled: list[str] = []
     pkgs_to_be_uninstalled: list[str] = []
     for pkg_name, pkg_info in dct.items():
         if not pkg_info[1] or all(pkg in dct for pkg in pkg_info[1]):
             pkgs_to_be_uninstalled.append(pkg_name)
         else:
             logger.info(
-                "Cannot uninstall %s. Required by: %s",
+                "Cannot uninstall %s: Required by %s",
+                pkg_name,
+                ", ".join(pkg_info[1].difference(dct)),
+            )
+            pkgs_to_not_be_uninstalled.append(pkg_name)
+
+    for pkg in pkgs_to_not_be_uninstalled:
+        del dct[pkg]
+    pkgs_to_be_uninstalled.clear()
+
+    for pkg_name, pkg_info in dct.items():
+        if not pkg_info[1] or all(pkg in dct for pkg in pkg_info[1]):
+            pkgs_to_be_uninstalled.append(pkg_name)
+        else:
+            logger.info(
+                "Cannot uninstall %s: Required by %s",
                 pkg_name,
                 ", ".join(pkg_info[1].difference(dct)),
             )
 
     if pkgs_to_be_uninstalled:
         logger.info("Purging: %s", ", ".join(pkgs_to_be_uninstalled))
-        purge_packages(pkgs_to_be_uninstalled, uninstall_args)
+        uninstall_packages(pkgs_to_be_uninstalled, uninstall_args)
     return 0
